+++ CHANGES.rst
@@ -47,6 +47,10 @@ Changes:
   generate ``$schema`` definition and return reference to expected/provided schema in responses
   (relates to `#157 <https://github.com/crim-ca/weaver/issues/157>`_)
   Only utilities are added, not all routes provide the information yet.
+- Add validation of ``schema`` field under ``Format`` schema (as per `opengeospatial/ogcapi-processes schema format.yml
+  <https://github.com/opengeospatial/ogcapi-processes/blob/master/core/openapi/schemas/format.yaml>`_) such that only
+  URL formatted strings are allowed, or alternatively and explicit JSON definition. Previous definitions that would
+  indicate an empty string schema are dropped since ``schema`` is optional.
 
 Fixes:
 ------